refactor: move implementations from controller to endpoint class
